pyproject.toml
[project]
name = "free-interviewcoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["python-dotenv>=1.1.0", "supabase>=2.15.0", "rich>=14.0.0"]

[dependency-groups]
dev = ["ipykernel>=6.29.5"]
